Fix schema operation replay. (#13)
Fix the bug in the serialization of schema operations. Since the schema operations can contain multiple messages now, the number of messages should also be serialized, which is missed, casuing the deserializing can get wrong value.

Fix the bug in parsing from rocksdb when starting. It's caused by wrongly using a RepeatedPtrField to parse all transactions information and mix log of different transactions into a single message.

Fix the bug in calculating the latest transaction number (txn). Currently, the txn of schema and range log is missed when calculating the latest txn when starting, making it possible that logservice returns a too small txn to txservice and a same txn is used twice, causing conflicts.

Also throw rocksdb error to avoid log service hang.